% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bols_post_ee.R
\name{bols_post_ee}
\alias{bols_post_ee}
\title{Posterior distribution-Gaussian Bolstering error estimation}
\usage{
bols_post_ee(
  psi,
  X,
  Y,
  S,
  mc_sample = 100,
  degree = 1,
  mod = NULL,
  loss = quad_loss
)
}
\arguments{
\item{psi}{Regression function}

\item{X}{X data}

\item{Y}{Y data}

\item{S}{List of kernels. If list size is 1, then use the same kernel for all points}

\item{mc_sample}{Number of samples for Monte Carlo integration}

\item{degree}{Degree of the polynomial for Bayesian regression}

\item{mod}{Object with the fitted Bayesian Regression. If not provided, the respective model will be fitted}

\item{loss}{A loss function}
}
\value{
The posterior distribution-Gaussian Bolstering error estimation considering Monte Carlo integration
}
\description{
Compute the posterior distribution-Gaussian bolstering error estimator in regression. The posterior distribution is given by Bayesian regression.
}
\examples{
psi_star <- function(x){(1 + sum(x))^2}
X <- matrix(runif(2*20),nrow = 20,ncol = 2)
Y <- apply(X,1,psi_star) + rnorm(20,sd = 0.25)
mod <- lm(Y ~ .,data.frame(poly(X,2,raw = TRUE),Y))
psi <- function(x){predict(mod,data.frame(poly(rbind(x),2,raw = TRUE)))}
S <- kernel_estimator(X)
bols_post_ee(psi,X,Y,mc_sample = 100,degree = 1,mod = NULL,S = S)
}
\references{
D. Marcondes, U. Braga-Neto. Generalized Resubstitution for Error Estimation in Regression. 2024.
}
