% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolstering.R
\name{bolstering}
\alias{bolstering}
\title{Gaussian Bolstering error estimation}
\usage{
bolstering(psi, X, Y, S, mc_sample = 100, loss = quad_loss)
}
\arguments{
\item{psi}{Estimated function}

\item{X}{X data}

\item{Y}{Y data}

\item{S}{List of kernels. If list size is 1, then use the same kernel for all points}

\item{mc_sample}{Number of samples for Monte Carlo integration}

\item{loss}{A loss function}
}
\value{
The Gaussian bolstering error estimator considering the Monte Carlo integration
}
\description{
Compute the Gaussian bolstering error estimator in regression. If the kernel dimension is d + 1 the smoothing is on both directions.
}
\examples{
psi_star <- function(x){(1 + sum(x))^2}
X <- matrix(runif(2*20),nrow = 20,ncol = 2)
Y <- apply(X,1,psi_star) + rnorm(20,sd = 0.25)
mod <- lm(Y ~ .,data.frame(poly(X,2,raw = TRUE),Y))
psi <- function(x){predict(mod,data.frame(poly(rbind(x),2,raw = TRUE)))}
S <- kernel_estimator(cbind(X,Y),method = "mpe")
bolstering(psi,X,Y,S)
}
\references{
D. Marcondes, U. Braga-Neto. Generalized Resubstitution for Error Estimation in Regression. 2024.
}
