% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ee.R
\name{cv_ee}
\alias{cv_ee}
\title{Cross-validation error estimator}
\usage{
cv_ee(X, Y, k = 10, model, loss = quad_loss)
}
\arguments{
\item{X}{X data}

\item{Y}{Y data}

\item{k}{Number of folds}

\item{model}{A function that fit a model and return its prediction function psi}

\item{loss}{A loss function}
}
\value{
The cross-validation error estimation
}
\description{
Compute the cross-validation error estimator
}
\examples{
psi_star <- function(x){(1 + sum(x))^2}
X <- matrix(runif(2*20),nrow = 20,ncol = 2)
Y <- apply(X,1,psi_star) + rnorm(20,sd = 0.25)
model <- function(X,Y){
  mod_pol <- lm(Y ~ .,data.frame(poly(X,2,raw = TRUE),Y))
  psi <- function(x){predict(mod_pol,data.frame(poly(rbind(x),2,raw = TRUE)))}
  return(psi)
}
cv_ee(X,Y,k = 10,model)
}
