% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mahalanobis.R
\name{dist_mahalanobis}
\alias{dist_mahalanobis}
\title{Mahalanobis distance}
\usage{
dist_mahalanobis(X, S = NULL)
}
\arguments{
\item{X}{X data}

\item{S}{Symmetric and positive definite matrix. If NULL the identity matrix is considered}
}
\value{
A Mahalanobis distance matrix
}
\description{
Mahalanobis distance matrix between sample points
}
\examples{
X <- matrix(runif(2*5),nrow = 5,ncol = 2)
dist_mahalanobis(X,rbind(c(1,0.5),c(0.5,1)))
}
