% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_estimator.R
\name{kernel_estimator}
\alias{kernel_estimator}
\title{Estimate the Kernel in Gaussian Bolstering}
\usage{
kernel_estimator(
  X,
  method = "chi",
  S = NULL,
  S0 = NULL,
  mc_sample = 100,
  ec = 1e-08,
  grid_delta = 0.001,
  lambda = 1,
  trace = T
)
}
\arguments{
\item{X}{X data}

\item{method}{Estimation method. Should be "mm" for method of moments, "chi" for chi approximation and "mpe"for maximum pseudolikelihood estimation}

\item{S}{Symmetric and positive definite matrix $\Sigma$ for the method of moments estimator. If NULL the identity matrix is considered}

\item{S0}{List of initial kernels for the EM algorithm of the MPE. If NULL the identity matrix is considered}

\item{mc_sample}{Number of samples for Monte Carlo integration}

\item{ec}{Stop criteria of the EM algorithm}

\item{grid_delta}{Step of grid search for method of moments estimation}

\item{lambda}{Parameter of EM algorithm}

\item{trace}{Trace EM algorithm}
}
\value{
A list of kernels
}
\description{
Estimate the bolster kernel via the maximum pseudolikelihood estimator (MPE) or via the Method of Moments, through Monte Carlo integration (exact) or
the chi approximation
}
\examples{
X <- matrix(runif(2*100),nrow = 100,ncol = 2)
K <- kernel_estimator(X,method = "mpe")
}
\references{
D. Marcondes, U. Braga-Neto. Generalized Resubstitution for Error Estimation in Regression. 2024.
}
