% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_ee.R
\name{posterior_ee}
\alias{posterior_ee}
\title{Posterior distribution generalized resubstitution error}
\usage{
posterior_ee(
  psi,
  X,
  Y,
  mc_sample = 100,
  degree = 1,
  mod = NULL,
  loss = quad_loss
)
}
\arguments{
\item{psi}{Regression function}

\item{X}{X data}

\item{Y}{Y data}

\item{mc_sample}{Number of samples for Monte Carlo integration}

\item{degree}{Degree of the polynomial for Bayesian regression}

\item{mod}{Object with the fitted Bayesian Regression. If not provided, the model will be fitted}

\item{loss}{A loss function}
}
\value{
The posterior probability error estimation considering the Monte Carlo mean loss under the posterior distribution
}
\description{
Compute the posterior distribution generalized resubstitution error in regression with posterior distribution given by Bayesian regression.
}
\examples{
psi_star <- function(x){(1 + sum(x))^2}
X <- matrix(runif(2*20),nrow = 20,ncol = 2)
Y <- apply(X,1,psi_star) + rnorm(20,sd = 0.25)
mod_pol <- lm(Y ~ .,data.frame(poly(X,2,raw = TRUE),Y))
psi <- function(x){predict(mod_pol,data.frame(poly(rbind(x),2,raw = TRUE)))}
posterior_ee(psi,X,Y,mc_sample = 100,degree = 2,mod = NULL)
}
\references{
D. Marcondes, U. Braga-Neto. Generalized Resubstitution for Error Estimation in Regression. 2024.
}
