% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{quad_loss}
\alias{quad_loss}
\title{Loss functions}
\usage{
quad_loss(obs, pred)
}
\arguments{
\item{obs}{Vector with observed data}

\item{pred}{Vector with predicted data}
}
\value{
A vector of losses.
}
\description{
Common loss functions for learning problems. \emph{quad_loss} for quadratic loss,
\emph{simple_loss} for the simple loss function and \emp{mae_loss} for the mean absolute error (MAE) loss.
}
